pyproject.toml
[tool.poetry]
name = "postgis-era5"
version = "0.1.0"
description = ""
authors = ["Jeffrey Tsang <jtsang93@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.4.0"
geopandas = "^0.10.2"
Shapely = "^1.8.0"
SQLAlchemy = "^1.4.31"
GeoAlchemy2 = "^0.10.2"
psycopg2-binary = "^2.9.3"
xarray = "^0.21.1"
netCDF4 = "^1.5.8"
isort = "^5.10.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
